make api calls only once in app.jsx
