+++ client/src/Pages/News.jsx
@@ -1,24 +1,10 @@
 import React, { useEffect, useState } from 'react';
 import NewsCard from '../components/NewsCard';
-import { fetchAPIData } from '../utils/fetchAPIData.js';
 
+const News = (props) => {
+    
+    const news = props.news || []; 
 
-const News = () => {
-    
-    
-    const [news, setNews] = useState([]);
-
-    useEffect(() => {
-        const getNews = async () => {
-            const data = await fetchAPIData('news');
-            setNews(data);
-        };
-
-        getNews();
-        
-    }, []);
-    
-    
     return (
         <div className='news'>
             <h1>All News</h1>
