+++ client/src/app.jsx
@@ -1,4 +1,4 @@
-import React from 'react'
+import React, {useState, useEffect} from 'react'
 
 import NavBar from './components/NavBar'
 import About from './Pages/About'
@@ -10,6 +10,7 @@ import ArtistCardLarge from './components/ArtistCardLarge'
 import Shop from './Pages/Shop'
 import ContactForm from './Pages/Contact'
 import Footer from './components/Footer'
+import { fetchAPIData } from './utils/fetchAPIData.js';
 
 import {
     BrowserRouter as Router,
@@ -18,6 +19,31 @@ import {
 } from "react-router-dom";
 
 const App = () => {
+
+
+    const [artists, setArtists] = useState([]);
+    const [news, setNews] = useState([]);
+  
+    useEffect(() => { 
+        const getArtists = async () => {
+            const data = await fetchAPIData('artists'); // make api calls just once
+            setArtists(data);
+        };
+
+        getArtists();
+    }, []);
+
+    useEffect(() => {
+        const getNews = async () => {
+            const data = await fetchAPIData('news');
+            setNews(data);
+        };
+
+        getNews();
+        
+    }, []);
+
+
     return (
         <div>
             <Router>
@@ -26,16 +52,16 @@ const App = () => {
                     <Route path="/" element={
                         <>
                             <About />
-                            <NewsBanner />
-                            <Artists />
+                            <NewsBanner news={news}/>
+                            <Artists artists={artists}/>
                             <ContactForm />
                             <Shop />
                         </>
                         } />
                     <Route path="/About" element={<About />} />
-                    <Route path="/News" element={<News />} />
+                    <Route path="/News" element={<News news={news}/>} />
                     <Route path="/News/:title" element={<NewsCardLarge/>} />
-                    <Route path="/Artists" element={<Artists />} />
+                    <Route path="/Artists" element={<Artists artists={artists}/>} />
                     <Route path="/Artists/:name" element={<ArtistCardLarge />} />
                     <Route path="/Shop" element={<Shop />} />
                     <Route path="/Contact" element={<ContactForm />} />
